Add SSL certification expiration monitoring feature
Implemented a feature to automatically check the expiration status of SSL certificates for multiple specified domains. Created new `template.md` and `domains.php` files to store the formatted output and the list of domains to be checked, respectively. Updated the `README.md` and `check.php` files to accommodate the new feature. Also, added a new workflow that triggers a PHP script to run the check daily, and commit any changes to the `README.md`.