Add date-time stamp to SSL expiration monitoring report
A line has been added to the SSL expiration monitoring report, providing a date-time stamp of when the report was generated. This was done by modifying the template.md and check.php files. The PHP script in check.php now includes a line that formats the date and time at which the SSL certificate check was performed. This added information will be helpful for users to understand when the SSL certificate details were last checked.