Update workflow to send email notification & modify certificate check
This commit adds email notifications to the workflow when the SSL certificate has expired, utilizing dawidd6/action-send-mail action. This is achieved by checking the output of check.php and sending an email if it contains 'has expired'.

In addition to this, the behavior of check.php has been modified. Now, the details are echoed only if the certificate will expire in less than 25 days. If the certificate has expired, the output contains a special warning message. The exception message has also been updated to be more explicit, adding that the domain in question may have expired.