Refactor SSL monitoring and embed file content into the code
The SSL monitoring script was significantly refactored for simplicity and maintainability.
Previously, domain information was retrieved from a separate 'domain.php' file, and email report template was read from 'template.md'. Having these as physically separate entities added some complexity to the code and made it less easy to manage. Now, these data are embedded directly into the script ('check.php') as a constant and an environmental variable respectively. It makes code more streamlined and easier to manage.

Also, updated the Github workflow file to adjust for the removal of 'domains.php' and add a condition to check if an SSL certificate is expired. This is to further refine the alert system, allowing for more granular and relevant alerts based on the certificate status.